Migrating dev accessibility changes to staging
Fixed AccessibilityEvent type in SubmisstionTestResultFragment so that the result card is focused and therefore announced.
Changed the xml attributes of the result card so that it can be focused with an id and gave it appropriate accessibility attributes for announcement.
added correct German translation for button
marked importantForAccesibility to no for TAN so that the double tap to activate speech is not spoken.
Closes re-opened accessibility JIRA tasks: EXPOSUREAPP-1540, EXPOSUREAPP-1541, EXPOSUREAPP-1542