Never report CancellationException (#1589)
There is a number of place in the app with the following code pattern:

try {
    // do some suspending operation
} catch(e: Exception) {
    e.report(ExceptionCategory.EXPOSURENOTIFICATION) // or similar
}

In all such place if the corresponding suspending gets cancelled for any reason, the resulting CancellationException will get reported as an application crash. This fix prevents it from happening, since a cancellation exception is a part of a normal operation, not a failure that should be reported.