Adopt TRL encoding to ENF restrictions (EXPOSUREAPP-4039) (#1763)
* Adopt Exposure Window test cases to new TRL encoding

keys with report type RECURSIVE are dropped by Android in Exposure
Window mode. The new encoding uses RECUSRIVE only with TRL 1 and 2
which are dropped anyway.

https://jira-ibs.wbs.net.sap/browse/EXPOSUREAPP-4039

* Adopt default App Config to new TRL encoding

keys with report type RECURSIVE are dropped by Android in Exposure
Window mode. The new encoding uses RECUSRIVE only with TRL 1 and 2
which are dropped anyway.

https://jira-ibs.wbs.net.sap/browse/EXPOSUREAPP-4039
https://github.com/corona-warn-app/cwa-protocol-buffers/commit/634c13c687a8f84d737a23c84e0e7c86e8e8bf9d

Co-authored-by: BMItter <46747780+BMItter@users.noreply.github.com>