Change risk_card_body_saved_days_full text to prevent misunderstanding
The text "Risiko-Ermittlung dauerhaft aktiv" is mesleading. It sounds like the app is up and running and scans for other users, but in fact it is only active, if GPS is enabled. which is not clearly displayed. (only hidden behind the explanation `settings_tracing_status_location_body`, whichnot everyone would look at, if everything is green and the message displays "dauerhaft aktiv")