Adapt the rest of the app to different base45 decoding behaviour (failure value instead of thrown exceptions)
The main part of the change is, that in the VaccinationContainer the contents of the certificate are optional (filled only when decoding was successful).
Only when the container is converted `toVaccinationCertificate()` in the `VaccinationCertificate` those fields are always present.

This change makes it so that vaccination certificates with invalid Base45 are silently removed, if they were imported in an earlier version of the app.